+++ src/Symfony/Bridge/Monolog/Processor/RouteProcessor.php
@@ -13,7 +13,7 @@ namespace Symfony\Bridge\Monolog\Processor;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\FinishRequestEvent;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 use Symfony\Contracts\Service\ResetInterface;
 
@@ -49,7 +49,7 @@ class RouteProcessor implements EventSubscriberInterface, ResetInterface
         $this->routeData = [];
     }
 
-    public function addRouteData(GetResponseEvent $event)
+    public function addRouteData(RequestEvent $event)
     {
         if ($event->isMasterRequest()) {
             $this->reset();