+++ src/Symfony/Component/HttpKernel/DataCollector/RouterDataCollector.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\HttpKernel\DataCollector;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
+use Symfony\Component\HttpKernel\Event\ControllerEvent;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
@@ -69,7 +69,7 @@ class RouterDataCollector extends DataCollector
      *
      * @final since Symfony 4.3
      */
-    public function onKernelController(FilterControllerEvent $event)
+    public function onKernelController(ControllerEvent $event)
     {
         $this->controllers[$event->getRequest()] = $event->getController();
     }