+++ src/Symfony/Component/HttpKernel/Event/RequestEvent.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\HttpKernel\Event;
 
+use Symfony\Component\HttpFoundation\Response;
+
 /**
  * Allows to create a response for a request.
  *
@@ -20,6 +22,37 @@ namespace Symfony\Component\HttpKernel\Event;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class RequestEvent extends GetResponseEvent
+class RequestEvent extends KernelEvent
 {
+    private $response;
+
+    /**
+     * Returns the response object.
+     *
+     * @return Response|null
+     */
+    public function getResponse()
+    {
+        return $this->response;
+    }
+
+    /**
+     * Sets a response and stops event propagation.
+     */
+    public function setResponse(Response $response)
+    {
+        $this->response = $response;
+
+        $this->stopPropagation();
+    }
+
+    /**
+     * Returns whether a response was set.
+     *
+     * @return bool Whether a response was set
+     */
+    public function hasResponse()
+    {
+        return null !== $this->response;
+    }
 }